requirements.txt
Flask>=2.0.0
Werkzeug>=2.0.0
openai
azure-identity
python-dotenv
azure-cosmos
pydantic
